Added new authentication method of bearer with MCP client
