+++ src/mcp_server/__init__.py
@@ -5,7 +5,7 @@ from typing import Literal
 import click
 from dotenv import load_dotenv
 
-from mcp_server.lib import Logger
+from mcp_server.lib.log import Logger
 
 from .server_sse import serve as serve_sse
 from .server_stdio import serve as serve_stdio
@@ -17,7 +17,7 @@ load_dotenv()
 @click.option("--host", default="localhost", help="Host to bind to")
 @click.option("--port", default=8000, help="Port to listen on")
 @click.option("--endpoint", "-e", help="Pagoda endpoint URL", required=True)
-@click.option("--token", "-t", help="Pagoda user token", required=True)
+@click.option("--token", "-t", help="Pagoda user token", required=False)
 @click.option(
     "--loglevel",
     "-l",
@@ -30,8 +30,14 @@ load_dotenv()
 @click.option(
     "--transport",
     default="sse",
-    type=click.Choice(["sse", "streamable-http", "stdio"]),
-    help="Transport protocol to use ('sse' or 'streamable-http')",
+    type=click.Choice(["sse", "stdio"]),
+    help="Transport protocol to use ('sse' or 'stdio')",
+)
+@click.option(
+    "--auth",
+    default="bearer",
+    type=click.Choice(["bearer", "azure"]),
+    help="Authentication method to use ('bearer' or 'azure')",
 )
 def main(
     host: str,
@@ -40,6 +46,7 @@ def main(
     token: str,
     loglevel: str,
     transport: Literal["sse", "stdio"],
+    auth: Literal["bearer", "azure"],
 ) -> None:
     Logger.setLevel(logging.getLevelName(loglevel))
 
@@ -48,7 +55,7 @@ def main(
             asyncio.run(serve_stdio(endpoint, token))
 
         case "sse":
-            asyncio.run(serve_sse(host, port, endpoint, token))
+            asyncio.run(serve_sse(host, port, auth, endpoint, token))
 
 
 if __name__ == "__main__":
